Editor interface + New texture system (#2)
* Add ImGui support for frame buffers.

* Starting ImGUI viewport render.

* Working viewport.

* Add docking support.

* Starting GameObject explorer.

* Add component register and component editors.

* Add component register and component editors.

* Add parallax mapping. Fix normal mapping.

* Add new texture engine.

* Add new texture engine. Fix directional light.

+++ src/vulkan/VKModel.cpp
@@ -13,7 +13,7 @@ void VKModel::reinitializeBuffer() {
     _instancingBuffer = std::make_unique<StagingBuffer>(
             _vkApplication,
             VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            _instancingStructSize * BUFFER_DEFAULT_SIZE
+            static_cast<uint32_t>(_instancingStructSize) * BUFFER_DEFAULT_SIZE
     );
     _data.resize(_instancingStructSize * BUFFER_DEFAULT_SIZE, 0);
 }
@@ -27,7 +27,7 @@ VKModel::VKModel(Application* application, std::vector<VKMesh*> meshes) :
         _instancingBuffer(std::make_unique<StagingBuffer>(
                 _vkApplication,
                 VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-                _instancingStructSize * BUFFER_DEFAULT_SIZE
+                static_cast<uint32_t>(_instancingStructSize) * BUFFER_DEFAULT_SIZE
         )),
         _data(_instancingStructSize * BUFFER_DEFAULT_SIZE, 0),
         _dataChangeRange(0, 0) {
